% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVRanges.R
\docType{methods}
\name{annotation,MVRanges-method}
\alias{annotation,MVRanges-method}
\title{simple annotations of variants (using TxDb.Hsapiens.NCBI.rCRS)}
\usage{
\S4method{annotation}{MVRanges}(object)
}
\arguments{
\item{object}{an MVRanges}
}
\value{
an MVRanges, with annotations from TxDb.Hsapiens.NCBI.rCRS
}
\description{
nb. Xiaowu's consequences are Synonymous, Missense, LoF, rRNA, tRNA, D-loop
so we kind of have to have D-loop annotation to tabulate these.  In order to
avoid doing this constantly, we cache the annotation in the MVRange's object
metadata() list.
}
\details{
for reference, anno_rCRS was prepared as follows:

columns <- c("TXNAME","GENEID")
  anno <- transcripts(TxDb.Hsapiens.NCBI.rCRS, columns=columns)
  seqlevels(anno) <- "chrM" # just in case 
  anno$GENEID <- unlist(anno$GENEID) 
  is.na(anno[grep("id-", anno$TXNAME)]$GENEID) <- TRUE
  names(anno) <- ifelse(is.na(anno$GENEID), '', paste0("MT-", anno$GENEID))

# map gene names to expected symbols
  canon <- c("MT-COX1"="MT-CO1", "MT-COX2"="MT-CO2", "MT-COX3"="MT-CO3",
             "MT-CYTB"="MT-CYB")
  for (i in names(canon)) names(anno) <- sub(i, canon[i], names(anno))

# label the regions hit by various alterations 
  anno$region <- ifelse(!grepl("id-", anno$TXNAME), "coding",
                        ifelse(grepl("id-TRN", anno$TXNAME), "tRNA", "rRNA"))
  dLoop1 <- GRanges("chrM", IRanges(1, 576))
  seqinfo(dLoop1) <- seqinfo(anno)
  dLoop1$TXNAME <- NA_character_
  dLoop1$GENEID <- NA_character_
  dLoop1$region <- "D-loop"
  dLoop2 <- dLoop1
  end(dLoop2) <- 16569
  start(dLoop2) <- 16024
  anno <- c(dLoop1, anno, dLoop2)

predictCoding(MVRanges, TxDb, BSgenome) still requires the TxDb and BSgenome 
objects for rCRS, so this just saves us a little time in regenerating them.
}
