% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAlignments.R
\docType{methods}
\name{asBam,MAlignments-method}
\alias{asBam,MAlignments-method}
\title{fetch the filename for an MAlignments}
\usage{
\S4method{asBam}{MAlignments}(file)
}
\arguments{
\item{file}{an MAlignments}
}
\value{
a BAM filename (string)
}
\description{
fetch the filename for an MAlignments
}
