% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAlignmentsList.R
\docType{methods}
\name{asBam,MAlignmentsList-method}
\alias{asBam,MAlignmentsList-method}
\title{list the BAM files and their aligned genomes for MAlignmentsList elements}
\usage{
\S4method{asBam}{MAlignmentsList}(file)
}
\arguments{
\item{file}{an MAlignmentsList}
}
\value{
BAM file summary for the MAlignmentsList
}
\description{
(co-opting a generic from `Rsamtools`)
}
