% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignHaplogroup.R
\name{assignHaplogroup}
\alias{assignHaplogroup}
\title{placeholder for using fragment assignment to guess mitochondrial haplogroup}
\usage{
assignHaplogroup(mal, mincovg = 20)
}
\arguments{
\item{mal}{an MAlignments of mitochondrial reads}

\item{mincovg}{minimum coverage to bother proceeding (20, per Griffith 2014)}
}
\value{
a vector of haplogroup probabilities
}
\description{
placeholder for using fragment assignment to guess mitochondrial haplogroup
}
