% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAlignments.R
\docType{methods}
\name{BamViews,MAlignments-method}
\alias{BamViews,MAlignments-method}
\title{recreate the BamViews used to load an MAlignments}
\usage{
\S4method{BamViews}{MAlignments}(bamRanges)
}
\arguments{
\item{bamRanges}{an MAlignments}
}
\value{
a BamViews object
}
\description{
recreate the BamViews used to load an MAlignments
}
