% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callMT.R
\name{callMT}
\alias{callMT}
\title{call mitochondrial variants from the results of ATACseeker::getMT()}
\usage{
callMT(mtReads, p.lower = 0.1, p.error = 0.001, read.count = 2L, ...)
}
\arguments{
\item{mtReads}{mitochondrial reads, with bamViews, from getMT()}

\item{p.lower}{lower bound on binomial probability for a variant (0.1)}

\item{p.error}{error probability (influences the minimum VAF; 0.001)}

\item{read.count}{minimum read depth required to support a variant (2)}

\item{...}{any other arguments to pass (currently ignored)}
}
\description{
call mitochondrial variants against the reference sequence they aligned to
the appropriate cutoffs will depend on the estimated copy number of mtDNA
in the cell type assayed; leukocytes usually have between 100-500, while
hepatocytes can have thousands. Therefore different settings make sense
for different cell types. The defaults call a mutation with ~ 2% VAF.
}
