% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAlignments.R
\docType{methods}
\name{callVariants,MAlignments-method}
\alias{callVariants,MAlignments-method}
\title{call variants (shockingly enough)}
\usage{
\S4method{callVariants}{MAlignments}(x)
}
\arguments{
\item{x}{an MAlignments}
}
\value{
a VRanges object
}
\description{
call variants (shockingly enough)
}
