% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAlignmentsList.R
\docType{methods}
\name{callVariants,MAlignmentsList-method}
\alias{callVariants,MAlignmentsList-method}
\title{call variants (shockingly enough)}
\usage{
\S4method{callVariants}{MAlignmentsList}(x)
}
\arguments{
\item{x}{an MAlignmentsList}
}
\value{
a MVRangesList object
}
\description{
call variants (shockingly enough)
}
