% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectStats.R
\name{collectStats}
\alias{collectStats}
\title{collect various cross-correlation and window size stats for QC}
\usage{
collectStats(curbam, pe.param)
}
\arguments{
\item{curbam}{the name of the BAM file to do this for}

\item{pe.param}{paired end parameters, derived elsewhere}
}
\value{
the output of csaw::profileSites after massaging inputs
}
\description{
collect various cross-correlation and window size stats for QC
}
