% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVRangesList.R
\docType{methods}
\name{computeDeviations,MVRangesList,RangedSummarizedExperiment-method}
\alias{computeDeviations,MVRangesList,RangedSummarizedExperiment-method}
\title{compute deviations from expected counts for a list of genotyped specimens}
\usage{
\S4method{computeDeviations}{MVRangesList,RangedSummarizedExperiment}(object,
  annotations)
}
\arguments{
\item{object}{an MVRangesList with an RSE in metadata(object)$counts}

\item{annotations}{RangedSummarizedExperiment with motif matches for object}
}
\value{
a chromVARDeviations object
}
\description{
compute deviations from expected counts for a list of genotyped specimens
}
