% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVRangesList.R
\docType{methods}
\name{counts,MVRangesList-method}
\alias{counts,MVRangesList-method}
\title{retrieve counts}
\usage{
\S4method{counts}{MVRangesList}(object)
}
\arguments{
\item{object}{an MVRangesList}
}
\value{
counts (a RangedSummarizedExperiment)
}
\description{
retrieve counts
}
