% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVRangesList.R
\docType{methods}
\name{counts<-,MVRangesList,RangedSummarizedExperiment-method}
\alias{counts<-,MVRangesList,RangedSummarizedExperiment-method}
\title{load counts}
\usage{
\S4method{counts}{MVRangesList,RangedSummarizedExperiment}(object) <- value
}
\arguments{
\item{object}{an MVRangesList}

\item{value}{a RangedSummarizedExperiment with identical colnames}
}
\description{
load counts
}
