% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAlignments.R
\docType{methods}
\name{coverage,MAlignments-method}
\alias{coverage,MAlignments-method}
\title{estimated read coverage}
\usage{
\S4method{coverage}{MAlignments}(x)
}
\arguments{
\item{x}{an MAlignments}
}
\value{
estimated coverage (numeric)
}
\description{
estimated read coverage
}
