% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAlignmentsList.R
\docType{methods}
\name{coverage,MAlignmentsList-method}
\alias{coverage,MAlignmentsList-method}
\title{estimated read coverage for each element (cached on load)}
\usage{
\S4method{coverage}{MAlignmentsList}(x)
}
\arguments{
\item{x}{an MAlignmentsList}
}
\value{
estimated coverage (numeric vector)
}
\description{
(co-opting a generic from `IRanges`)
}
