% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRanges.R
\docType{methods}
\name{coverage,MRanges-method}
\alias{coverage,MRanges-method}
\title{estimated read coverage (recorded from the called MAlignments)}
\usage{
\S4method{coverage}{MRanges}(x)
}
\arguments{
\item{x}{an MRanges}
}
\value{
estimated coverage (numeric) from the called MAlignments
}
\description{
estimated read coverage (recorded from the called MAlignments)
}
