% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVRanges.R
\docType{methods}
\name{coverage,MVRanges-method}
\alias{coverage,MVRanges-method}
\title{estimated read coverage (recorded from the called MAlignments)}
\usage{
\S4method{coverage}{MVRanges}(x)
}
\arguments{
\item{x}{an MVRanges}
}
\value{
estimated coverage (numeric) from the called MAlignments
}
\description{
estimated read coverage (recorded from the called MAlignments)
}
