% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVRangesList.R
\docType{methods}
\name{coverage,MVRangesList-method}
\alias{coverage,MVRangesList-method}
\title{estimated read coverage (recorded from the called MAlignments)}
\usage{
\S4method{coverage}{MVRangesList}(x)
}
\arguments{
\item{x}{an MVRangesList}
}
\value{
estimated coverage from the called MAlignments[List]
}
\description{
estimated read coverage (recorded from the called MAlignments)
}
