% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVRangesList.R
\docType{methods}
\name{deviations,MVRangesList-method}
\alias{deviations,MVRangesList-method}
\title{retrieve deviations}
\usage{
\S4method{deviations}{MVRangesList}(object)
}
\arguments{
\item{object}{an MVRangesList}
}
\value{
deviations
}
\description{
retrieve deviations
}
