% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVRanges.R
\docType{methods}
\name{encoding,MVRanges-method}
\alias{encoding,MVRanges-method}
\title{simple helper to retrieve PASS'ing, protein-coding variants from MVRanges}
\usage{
\S4method{encoding}{MVRanges}(x)
}
\arguments{
\item{mvr}{an MVRanges}
}
\value{
subset of MVRanges passing filters within protein-coding regions
}
\description{
simple helper to retrieve PASS'ing, protein-coding variants from MVRanges
}
