% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVRangesList.R
\docType{methods}
\name{encoding,MVRangesList-method}
\alias{encoding,MVRangesList-method}
\title{simple helper to retrieve PASS'ing, coding variants from an MVRangesList}
\usage{
\S4method{encoding}{MVRangesList}(x)
}
\arguments{
\item{x}{an MVRangesList}
}
\value{
subsets of MVRangesList elements passing filters in coding regions
}
\description{
simple helper to retrieve PASS'ing, coding variants from an MVRangesList
}
