% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estK.R
\name{estK}
\alias{estK}
\title{Estimate a value of k to use with RUV}
\usage{
estK(dat, mod, method = "leek", vfilter = NULL, B = 20, seed = 777)
}
\arguments{
\item{dat}{A data matrix where loci are rows and samples are columns}

\item{mod}{A model matrix used to fit the data (variables of interest)}

\item{method}{The method used to estimate k ("leek" (default - asymptotic) or "be" (permutation-based))}

\item{vfilter}{Specify the number of loci to filter based on ranked variance across samples (default is to use all loci)}

\item{B}{Specify the number of permutations to use if method = "be" (default = 20)}

\item{seed}{Specify a random number for setting the seed if method = "be"}
}
\value{
An estimated value of k for using in RUVseq
}
\description{
This function is used to estimate a suitable value of k for use in RUV normalization
}
