% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAlignmentsList.R
\docType{methods}
\name{fileName,MAlignmentsList-method}
\alias{fileName,MAlignmentsList-method}
\title{list the BAM filenames for the elements of an MAlignmentsList}
\usage{
\S4method{fileName}{MAlignmentsList}(object)
}
\arguments{
\item{object}{an MAlignmentsList}
}
\value{
BAM file summary for the MAlignmentsList
}
\description{
list the BAM filenames for the elements of an MAlignmentsList
}
