% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragmentLengths.R
\name{fragmentLengths}
\alias{fragmentLengths}
\title{get fragment sizes for a GAlignmentPairs object (can be slow)}
\usage{
fragmentLengths(galp, plotAs = c("hist", "density", "none"), ...)
}
\arguments{
\item{galp}{the GAlignmentPairs object}

\item{plotAs}{hist[ogram, default], density, or none}

\item{...}{other arguments to pass along to density()}
}
\value{
(invisibly) the width of fragments in galp
}
\description{
get fragment sizes for a GAlignmentPairs object (can be slow)
}
