% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragPlot.R
\name{fragPlot}
\alias{fragPlot}
\title{Plot fragment size density for a set of samples.}
\usage{
fragPlot(frag.dists, label)
}
\arguments{
\item{frag.dists}{the computed densities of each fragment distribution}

\item{label}{the group label to be used in the title}
}
\description{
Plot fragment size density for a set of samples.
}
\seealso{
fragmentLengths
}

