% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get5primeCuts.R
\name{get5primeCuts}
\alias{get5primeCuts}
\title{get the 5' cut sites from a GAlignmentPairs (usually of an ATAC BAM)}
\usage{
get5primeCuts(galp, maxSize = 1000, shrink = TRUE, shrinkBy = NULL, ...)
}
\arguments{
\item{galp}{a GAlignmentPairs object}

\item{maxSize}{maximum insert size (for fragmentLengths(galp))}

\item{shrink}{move the reads around like in the original paper? (TRUE)}

\item{shrinkBy}{shrink by how much? (realistically this is pointless...)}
}
\value{
the 5' ends, from getEnds()
}
\description{
get the 5' cut sites from a GAlignmentPairs (usually of an ATAC BAM)
}
\examples{
library(Homo.sapiens)
CUX1_EGID <- org.Hs.egSYMBOL2EG[['CUX1']]
CUX1 <- reduce(transcriptsBy(Homo.sapiens, 'gene')[[CUX1_EGID]])
CUX1bump <- GRanges('chr7', IRanges(101499132, 101501052), '+')

}
