% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVRanges.R
\docType{methods}
\name{getAnnotations,MVRanges-method}
\alias{getAnnotations,MVRanges-method}
\title{simple helper to retrieve annotations (if present) from an MVRanges}
\usage{
\S4method{getAnnotations}{MVRanges}(annotations)
}
\arguments{
\item{mvr}{an MVRanges}
}
\value{
a GRanges of annotations, if present, or else NULL
}
\description{
simple helper to retrieve annotations (if present) from an MVRanges
}
