% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getComplexity.R
\name{getComplexity}
\alias{getComplexity}
\title{complexity estimates from a GAlignmentPairs object, formatted for preseqR}
\usage{
getComplexity(x)
}
\arguments{
\item{x}{the GAlp object or GRanges}
}
\value{
a matrix of estimates
}
\description{
complexity estimates from a GAlignmentPairs object, formatted for preseqR
}

