% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnds.R
\name{getEnds}
\alias{getEnds}
\title{convert a galp into a gr of 5' cuts}
\usage{
getEnds(gr)
}
\arguments{
\item{gr}{an intermediate GRanges from a GAlignmentPairs}
}
\value{
a gr of width-1 cuts
}
\description{
convert a galp into a gr of 5' cuts
}
