% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAlignmentsList.R
\docType{methods}
\name{getFragmentsPerSample,MAlignmentsList-method}
\alias{getFragmentsPerSample,MAlignmentsList-method}
\title{number of aligned fragments per element}
\usage{
\S4method{getFragmentsPerSample}{MAlignmentsList}(object)
}
\arguments{
\item{object}{an MAlignmentsList}
}
\value{
the fragments (length) for each element
}
\description{
(co-opting a generic from chromVAR to get reads per element)
}
