% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAlignmentsList.R
\docType{methods}
\name{getProperties,MAlignmentsList-method}
\alias{getProperties,MAlignmentsList-method}
\title{element properties for an MAlignmentsList}
\usage{
\S4method{getProperties}{MAlignmentsList}(ClassDef)
}
\arguments{
\item{ClassDef}{an MAlignmentsList}
}
\value{
properties for each element, as a DataFrame
}
\description{
(co-opting a defunct `methods` function back into a generic)
}
