% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQC.R
\name{getQC}
\alias{getQC}
\title{obtain QC information used by csaw from paired-end BAM files}
\usage{
getQC(name, pe.bam)
}
\arguments{
\item{name}{the name of the sample}

\item{pe.bam}{the name of the BAM file}
}
\value{
a list with components "qc" (metrics) and "frag.dist" (density)
}
\description{
obtain QC information used by csaw from paired-end BAM files
}
