% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapMT.R
\name{mapMT}
\alias{mapMT}
\title{map the mitochondrial reads to the revised Cambridge reference sequence}
\usage{
mapMT(mtReads, ...)
}
\arguments{
\item{mtReads}{mitochondrial reads as a GAlignments, typically from getMT()}

\item{...}{other arguments to be passed to gsnap}
}
\description{
map the mitochondrial reads to the revised Cambridge reference sequence
}
