% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAlignments.R
\docType{methods}
\name{metadata,MAlignments-method}
\alias{metadata,MAlignments-method}
\title{retrieve the rest of the stored information in an MAlignments object}
\usage{
\S4method{metadata}{MAlignments}(x)
}
\arguments{
\item{x}{an MAlignments}
}
\value{
a list of metadata (gr, bam, bai, readLength, genomeSize)
}
\description{
retrieve the rest of the stored information in an MAlignments object
}
