% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRanges.R
\name{MRanges}
\alias{MRanges}
\title{wrap a VRanges for mitochondrial use}
\usage{
MRanges(vr, coverage)
}
\arguments{
\item{vr}{the VRanges}

\item{covg}{estimated coverage}
}
\value{
an MRanges
}
\description{
wrap a VRanges for mitochondrial use
}
