% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVRanges.R
\docType{methods}
\name{MVRanges-methods}
\alias{MVRanges-methods}
\alias{coverage,MVRanges-method}
\alias{type,MVRanges-method}
\alias{pos,MVRanges-method}
\alias{show,MVRanges-method}
\alias{annotation,MVRanges-method}
\alias{getAnnotations,MVRanges-method}
\alias{encoding,MVRanges-method}
\alias{predictCoding,MVRanges,missing,missing,missing-method}
\alias{summarizeVariants,MVRanges,missing,missing-method}
\title{MVRanges methods (centralized).}
\usage{
\S4method{coverage}{MVRanges}(x)

\S4method{type}{MVRanges}(x)

\S4method{pos}{MVRanges}(x)

\S4method{show}{MVRanges}(object)

\S4method{annotation}{MVRanges}(object)

\S4method{getAnnotations}{MVRanges}(annotations)

\S4method{encoding}{MVRanges}(x)

\S4method{predictCoding}{MVRanges,missing,missing,missing}(query, subject,
  seqSource, varAllele, ...)

\S4method{summarizeVariants}{MVRanges,missing,missing}(query, subject, mode,
  ...)
}
\arguments{
\item{x}{an MVRanges}

\item{object}{an MVRanges}

\item{annotations}{an MVRanges}

\item{query}{an MVRanges}

\item{x}{an MVRanges}

\item{x}{an MVRanges}

\item{x}{an MVRanges}

\item{x}{an MVRanges}

\item{query}{an MVRanges, almost always after predictCoding(MVRanges)}
}
\value{
estimated coverage (numeric) from the called MAlignments

a vector of variant types ("SNV" or "indel")

a named CHARACTER vector of variant positions for MitImpact/HGVS
}
\description{
MVRanges methods (centralized).
}
