% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVRangesList.R
\docType{methods}
\name{MVRangesList-methods}
\alias{MVRangesList-methods}
\alias{coverage,MVRangesList-method}
\alias{counts<-,MVRangesList,RangedSummarizedExperiment-method}
\alias{counts,MVRangesList-method}
\alias{computeDeviations,MVRangesList,RangedSummarizedExperiment-method}
\alias{deviations,MVRangesList-method}
\alias{encoding,MVRangesList-method}
\alias{show,MVRangesList-method}
\title{MVRangesList methods (centralized).}
\usage{
\S4method{coverage}{MVRangesList}(x)

\S4method{counts}{MVRangesList,RangedSummarizedExperiment}(object) <- value

\S4method{counts}{MVRangesList}(object)

\S4method{computeDeviations}{MVRangesList,RangedSummarizedExperiment}(object,
  annotations)

\S4method{deviations}{MVRangesList}(object)

\S4method{encoding}{MVRangesList}(x)

\S4method{show}{MVRangesList}(object)
}
\arguments{
\item{x}{an MVRangesList}

\item{object}{an MVRangesList}

\item{value}{a RangedSummarizedExperiment with identical colnames}

\item{annotations}{RangedSummarizedExperiment with motif matches for object}

\item{object}{an MVRangesList}

\item{object}{an MVRangesList with an RSE in metadata(object)$counts}

\item{object}{an MVRangesList}

\item{x}{an MVRangesList}

\item{x}{an MVRangesList}
}
\value{
estimated coverage from the called MAlignments[List]
}
\description{
MVRangesList methods (centralized).
}
