% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVRangesList.R
\name{MVRangesList}
\alias{MVRangesList}
\title{wrap a VRangesList for mitochondrial use}
\usage{
MVRangesList(vrl)
}
\arguments{
\item{vrl}{the VRangesList}
}
\value{
an MVRangesList
}
\description{
wrap a VRangesList for mitochondrial use
}
