% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotComplexity.R
\name{plotComplexity}
\alias{plotComplexity}
\title{plot results from preseq (complexity estimation around features)}
\usage{
plotComplexity(x, ..., withCI = FALSE)
}
\arguments{
\item{x}{the thing whose complexity is to be plotted}
}
\value{
invisibly, the complexity estimates corresponding to the generated plot
}
\description{
FIXME: use the ... to allow for comparison of multiple libraries
}

