% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCoverage.R
\name{plotCoverage}
\alias{plotCoverage}
\title{utility fn from Patrick Abouyoun's 2010 seminar}
\usage{
plotCoverage(x, gr, col = "violet", xlab = "base", ylab = "reads",
  main = "Coverage", stranded = F, smoothen = 0, kf = "normal", ...)
}
\arguments{
\item{x}{a galp}

\item{gr}{a single GRanges range (i.e. seqname, start, end)}

\item{col}{what color to plot it}

\item{xlab}{x axis label for the plot}

\item{ylab}{y axis label for the plot}

\item{main}{main title for the plot}

\item{stranded}{plot stranded? (FALSE)}

\item{smoothen}{smooth the counts? (FALSE)}

\item{kf}{kernel function for smoothing (normal)}

\item{...}{arguments to pass along to coverage()}
}
\description{
utility fn from Patrick Abouyoun's 2010 seminar
}
\examples{
visualizing a specific enhancer:

SMOenh <- GRanges('chr6', IRanges(29744702, 29751934), '+')
seqinfo(SMOenh) <- seqinfo(Mus.musculus)[seqlevels(SMOenh)]

plotCoverage(galp, SMOenh)
plotCoverage(galp, SMOenh, stranded=TRUE)
plotCoverage(get5primeCuts(galp), SMOenh, stranded=TRUE)
plotCoverage(get5primeCuts(galp, shrink=FALSE), SMOenh, stranded=TRUE)

}
