% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStranded.R
\name{plotStranded}
\alias{plotStranded}
\title{convenience function}
\usage{
plotStranded(x, gr, ...)
}
\arguments{
\item{x}{coverage}

\item{gr}{a single GRange}

\item{...}{args to pass along to plotCoverage}
}
\description{
convenience function
}
