% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVRanges.R
\docType{methods}
\name{pos,MVRanges-method}
\alias{pos,MVRanges-method}
\title{positions of variants (for feeding to MitImpact and friends)}
\usage{
\S4method{pos}{MVRanges}(x)
}
\arguments{
\item{x}{an MVRanges}
}
\value{
a CHARACTER vector of variant position strings for MitImpact
}
\description{
positions of variants (for feeding to MitImpact and friends)
}
