% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVRanges.R
\docType{methods}
\name{predictCoding,MVRanges,missing,missing,missing-method}
\alias{predictCoding,MVRanges,missing,missing,missing-method}
\title{predict coding mutations (annotating the most likely consequences)}
\usage{
\S4method{predictCoding}{MVRanges,missing,missing,missing}(query, subject,
  seqSource, varAllele, ...)
}
\arguments{
\item{mvr}{an MVRanges}
}
\value{
an rCRS MVRanges with CONSEQUENCE, MTCSQ, and (REF/VAR)(CODON/AA)
}
\description{
Given an rCRS-situated set of variant calls, predict which may be coding,
and which are synonymous (so as to compute dN/dS and annotate regionally).
}
\details{
Like the other predictCoding() methods, one can expect mcols() containing:

- varAllele   (reverse complemented if the subject is on the negative strand)
- QUERYID     (usually blank; rCRS liftOver step can make indices confusing)
- TXID        (usually blank; each mitochondrial gene has one transcript)
- CDSID       (usually blank; each mitochondrial gene has one known CDS) 
- GENEID      (usually blank; mitochondrial genes are just ID'ed as 1-13)
- CDSLOC      
- PROTEINLOC
- CONSEQUENCE (synonymous/nonsynonymous/frameshift/nonsense/not translated)
- REFCODON
- VARCODON
- REFAA
- VARAA

Unlike the other predictCoding methods, many of these will usually be empty.
}
