% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processUcscCytoband.R
\name{processUcscCytoband}
\alias{processUcscCytoband}
\title{utility function to pull in UCSC cytoband / arm files 
this is then applied to the mcols() of a GRanges or GAlignmentPairs
and then one can trivially parallelize over cytobands or arms with split()}
\usage{
processUcscCytoband(filename)
}
\arguments{
\item{filename}{the UCSC cytoband file}
}
\value{
a GRanges with $chrArm and $chrBand as mcols
}
\description{
utility function to pull in UCSC cytoband / arm files 
this is then applied to the mcols() of a GRanges or GAlignmentPairs
and then one can trivially parallelize over cytobands or arms with split()
}
