% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCRS.R
\name{rCRS}
\alias{rCRS}
\title{liftOver and simplify an hg19 mitochondrial genome (variants) to rCRS}
\usage{
rCRS(mvr)
}
\arguments{
\item{mvr}{an MVRanges of variant calls}
}
\value{
an MVRanges of variant calls against rCRS
}
\description{
liftOver and simplify an hg19 mitochondrial genome (variants) to rCRS
}
