% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAlignments.R
\docType{methods}
\name{runLength,MAlignments-method}
\alias{runLength,MAlignments-method}
\title{median read length}
\usage{
\S4method{runLength}{MAlignments}(x)
}
\arguments{
\item{x}{an MAlignments}
}
\value{
read length
}
\description{
median read length
}
