% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAlignmentsList.R
\docType{methods}
\name{runLength,MAlignmentsList-method}
\alias{runLength,MAlignmentsList-method}
\title{estimated read length for each element (cached on load)}
\usage{
\S4method{runLength}{MAlignmentsList}(x)
}
\arguments{
\item{x}{an MAlignmentsList}
}
\value{
estimated coverage (numeric vector)
}
\description{
(co-opting a generic from `S4Vectors`)
}
