% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAlignments.R
\docType{methods}
\name{runValue,MAlignments-method}
\alias{runValue,MAlignments-method}
\title{genome size}
\usage{
\S4method{runValue}{MAlignments}(x)
}
\arguments{
\item{x}{an MAlignments}
}
\value{
genome size
}
\description{
genome size
}
