% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruvNorm.R
\name{ruvNorm}
\alias{ruvNorm}
\title{Remove unwanted variance (RUV)}
\usage{
ruvNorm(x, mod, cIdx, k = "estimate", scIdx, round = TRUE, epsilon = 1,
  tolerance = 1e-08, isLog = FALSE, ...)
}
\arguments{
\item{x}{A data matrix where loci are rows and samples are columns}

\item{mod}{A model matrix of variables to fit the data}

\item{cIdx}{A vector of loci/spike-ins in the data matrix to be used as negative controls (default = all loci)}

\item{k}{The number of factors of unwanted variation to be estimated/removed from the data (default = estimate)}

\item{scIdx}{A matrix of sample replicates to be used to estimate/remove unwanted variation}

\item{round}{Measures are rounded to form pseudo-counts}

\item{epsilon}{An offset to be added to avoid taking log of 0 (default = 1)}

\item{tolerance}{Selection tolerance for singular values to be considered positive (default = 1e-8)}

\item{isLog}{Whether the input matrix is already log-transformed (default = FALSE)}

\item{...}{Other arguments to be passed to estK}
}
\value{
A list object of the samples-by-factors matrix of estimated unwanted variation and the normalized counts (removed unwanted variation)
}
\description{
This function is a wrapper for RUVs from RUVSeq to remove unwanted variance using control genes.
Additionally, if the number of factors to remove is unknown, the value of k is estimated using
surrogate variable analysis. However, if the number of factors is known a priori, then k can be
specified.
}
