% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAlignments.R
\docType{methods}
\name{scanBamHeader,MAlignments-method}
\alias{scanBamHeader,MAlignments-method}
\title{fetch the header from an MAlignments' original BAM}
\usage{
\S4method{scanBamHeader}{MAlignments}(files)
}
\arguments{
\item{bamRanges}{an MAlignments}
}
\value{
a BAM header
}
\description{
fetch the header from an MAlignments' original BAM
}
