% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAlignments.R
\docType{methods}
\name{show,MAlignments-method}
\alias{show,MAlignments-method}
\title{display alignment records with read coverage estimate}
\usage{
\S4method{show}{MAlignments}(object)
}
\arguments{
\item{x}{an MAlignments}
}
\description{
display alignment records with read coverage estimate
}
