% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAlignmentsList.R
\docType{methods}
\name{show,MAlignmentsList-method}
\alias{show,MAlignmentsList-method}
\title{display alignment collections with element summaries}
\usage{
\S4method{show}{MAlignmentsList}(object)
}
\arguments{
\item{objects}{an MAlignmentsList}
}
\description{
display alignment collections with element summaries
}
