% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRanges.R
\docType{methods}
\name{show,MRanges-method}
\alias{show,MRanges-method}
\title{display variant calls with overall read coverage estimate}
\usage{
\S4method{show}{MRanges}(object)
}
\arguments{
\item{x}{an MRanges}
}
\description{
display variant calls with overall read coverage estimate
}
