% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVRanges.R
\docType{methods}
\name{show,MVRanges-method}
\alias{show,MVRanges-method}
\title{display variant calls with overall read coverage estimate}
\usage{
\S4method{show}{MVRanges}(object)
}
\arguments{
\item{x}{an MVRanges}
}
\description{
display variant calls with overall read coverage estimate
}
