% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVRangesList.R
\docType{methods}
\name{show,MVRangesList-method}
\alias{show,MVRangesList-method}
\title{display variant calls with overall read coverage estimate}
\usage{
\S4method{show}{MVRangesList}(object)
}
\arguments{
\item{x}{an MVRangesList}
}
\description{
display variant calls with overall read coverage estimate
}
