% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkPairs.R
\name{shrinkPairs}
\alias{shrinkPairs}
\title{This seems to work better than anything else for "shifting" cuts}
\usage{
shrinkPairs(x, shrinkBy)
}
\arguments{
\item{x}{a GAlignmentPairs or GRanges}

\item{shrinkBy}{shrink by this many bases}
}
\value{
the input object, widths shrunken by shrinkBy bases
}
\description{
This seems to work better than anything else for "shifting" cuts
}
