% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVRanges.R
\docType{methods}
\name{summarizeVariants,MVRanges,missing,missing-method}
\alias{summarizeVariants,MVRanges,missing,missing-method}
\title{summarize consensus pathogenicity estimates from MitImpact (2.0 or newer)}
\usage{
\S4method{summarizeVariants}{MVRanges,missing,missing}(query, subject, mode,
  ...)
}
\arguments{
\item{query}{an MVRanges, almost always after predictCoding(MVRanges)}
}
\value{
MitImpact output for variants where information is available
}
\description{
Note: this method requires an internet connection. For now, at least.
}
