% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAlignments.R
\docType{methods}
\name{Summary,MAlignments-method}
\alias{Summary,MAlignments-method}
\title{summary of object: reads, readLength, genomeSize, coverage}
\usage{
\S4method{Summary}{MAlignments}(x)
}
\arguments{
\item{x}{an MAlignments}
}
\value{
summary of x
}
\description{
specifically, 
c(length(x), runLength(x), seqlengths(x)[seqlevelsInUse(x)], coverage(x))
}
