% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAlignmentsList.R
\docType{methods}
\name{Summary,MAlignmentsList-method}
\alias{Summary,MAlignmentsList-method}
\title{summary of an MAlignmentsList}
\usage{
\S4method{Summary}{MAlignmentsList}(x)
}
\arguments{
\item{x}{an MAlignmentsList}
}
\value{
a DataFrame
}
\description{
(generic defined in base, no less)
}
