% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tssQcPlot.R
\name{tssQcPlot}
\alias{tssQcPlot}
\title{what you would expect.  not terribly novel; maybe use esATAC's instead?
FIXME: generalize this to handle CTCF/TSS plots, etc. according to genome(x)}
\usage{
tssQcPlot(x, anno, ordering, span = 5000, smoothing = 50, ...)
}
\arguments{
\item{x}{a GAlignmentPairs or GRanges to be scored w/featureScores}

\item{anno}{annotations, e.g. Homo.sapiens or Mus.musculus}

\item{ordering}{arg to pass to binPlots()}

\item{span}{smoother span}

\item{smoothing}{smoothing width}

\item{...}{args to pass along to binPlots}
}
\value{
whatever binPlots returns
}
\description{
what you would expect.  not terribly novel; maybe use esATAC's instead?
FIXME: generalize this to handle CTCF/TSS plots, etc. according to genome(x)
}
