% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVRanges.R
\docType{methods}
\name{type,MVRanges-method}
\alias{type,MVRanges-method}
\title{types of variants (for feeding to MitImpact and friends)}
\usage{
\S4method{type}{MVRanges}(x)
}
\arguments{
\item{x}{an MVRanges}
}
\value{
a vector of variant types ("SNV" or "indel")
}
\description{
types of variants (for feeding to MitImpact and friends)
}
