% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAlignmentsList.R
\docType{methods}
\name{updateObject,MAlignmentsList-method}
\alias{updateObject,MAlignmentsList-method}
\title{cache summary stats and perform any necessary updates for an MAlignmentsList}
\usage{
\S4method{updateObject}{MAlignmentsList}(object)
}
\arguments{
\item{object}{an MAlignmentsList, usually lacking some cached information}
}
\value{
an updated MAlignmentsList with cached metadata summaries
}
\description{
cache summary stats and perform any necessary updates for an MAlignmentsList
}
