% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAlignments.R
\docType{methods}
\name{Views,MAlignments-method}
\alias{Views,MAlignments-method}
\title{recreate the BamViews needed to call variants}
\usage{
\S4method{Views}{MAlignments}(subject)
}
\arguments{
\item{subject}{an MAlignments}
}
\value{
a BamViews object
}
\description{
recreate the BamViews needed to call variants
}
