% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAlignments.R
\docType{methods}
\name{yieldSize,MAlignments-method}
\alias{yieldSize,MAlignments-method}
\title{genome size}
\usage{
\S4method{yieldSize}{MAlignments}(object)
}
\arguments{
\item{x}{an MAlignments}
}
\value{
genome size
}
\description{
genome size
}
