% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atacPairedEnd.R
\name{atacPairedEnd}
\alias{atacPairedEnd}
\title{pull in (filter and assemble as an object) some or all of the "useful" reads
in a paired-end ATACseq run (i.e., the majority of ATACseq studies)}
\usage{
atacPairedEnd(bam, bamParams = NULL, which = NULL, ...)
}
\arguments{
\item{bam}{character string, the BAM file to parse}

\item{bamParams}{optional parameters to pass through to ScanBamParam (none)}

\item{which}{optional GRanges with specific regions to extract (all)}
}
\value{
a GenomicAlignmentPairs object
}
\description{
pull in (filter and assemble as an object) some or all of the "useful" reads
in a paired-end ATACseq run (i.e., the majority of ATACseq studies)
}
\examples{
\dontrun{

data(cytobands_hg19)
library(Homo.sapiens)
del7q <- byBand(cytobands_hg19)$chr7q22.1
galp <- atacPairedEnd('chr7.q10.CD49f_r1.hg19.bam', which=del7q)

library(Mus.musculus)
chr6 <- GRanges('chr6', IRanges(1, seqlengths(Mus.musculus)['chr6']), '*')
galp <- atacPairedEnd("A2.mm10.unique.bam",
                      bamParams=properPairedEndAtacFilters(which=chr6))

}

}
