% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callMT.R
\name{callMT}
\alias{callMT}
\title{call mitochondrial variants from an MAlignments object}
\usage{
callMT(mal, p.lower = 0.1, read.count = 2L, total.count = 10L,
  rCRS = FALSE, parallel = FALSE)
}
\arguments{
\item{mal}{an MAlignments (or, potentially, an MAlignmentsList)}

\item{p.lower}{lower bound on binomial probability for a variant (0.1)}

\item{read.count}{minimum alt read depth required to support a variant (2)}

\item{total.count}{minimum total read depth required to keep a variant (10)}

\item{rCRS}{lift to rCRS if not already hg38/GRCh38? (FALSE)}

\item{parallel}{try to run in parallel? (FALSE)}
}
\description{
FIXME: figure out a way to reprocess extracted chrM/MT reads against rCRS,
       regardless of what reference they were originally aligned against.
}
