% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAlignments.R
\docType{methods}
\name{callVariants,MAlignments-method}
\alias{callVariants,MAlignments-method}
\title{call variants (shockingly enough)}
\usage{
\S4method{callVariants}{MAlignments}(x, p.lower = 0.1, p.error = 0.001,
  read.count = 10)
}
\arguments{
\item{x}{an MAlignments}

\item{p.lower}{lower bound on variant probability to call (0.1)}

\item{p.error}{estimated error probability for a variant (0.001)}

\item{read.count}{minimum number of reads required to support a variant (10)}
}
\value{
a VRanges object
}
\description{
call variants (shockingly enough)
}
