% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEsts.R
\name{getEsts}
\alias{getEsts}
\title{Estimate the complexity of a library or sample based on unique fragments
using Daley & Smith's implementation of Good & Toulmin's rational function
approximation to solve the missing species problem.}
\usage{
getEsts(xx, withCI = FALSE, ...)
}
\arguments{
\item{xx}{The fragments or sample of fragments}

\item{withCI}{Have preseq compute 95% confidence intervals for plotting?}

\item{...}{Other arguments to pass on to preseq}
}
\value{
A data frame with results
}
\description{
Estimate the complexity of a library or sample based on unique fragments
using Daley & Smith's implementation of Good & Toulmin's rational function
approximation to solve the missing species problem.
}

