% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEsts.R
\name{getEsts}
\alias{getEsts}
\title{Complexity estimation for ATACseq libraries.}
\usage{
getEsts(xx, withCI = FALSE, ...)
}
\arguments{
\item{xx}{The fragments or sample of fragments}

\item{withCI}{Have preseq compute 95 percent confidence intervals for plots?}

\item{...}{Other arguments to pass on to preseq}
}
\value{
A data frame with results
}
\description{
Estimate the complexity of a library or sample based on unique fragments
using Daley and Smith's implementation of Good-Toulmin rational function
approximation to solve the missing species problem.
}
