% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMT.R
\name{getMT}
\alias{getMT}
\title{grab the mitochondrial reads from a BAM and estimate their fraction
nb. this could probably be done faster for a list of BAMs but it's not
nb. nb. returns NuMt-depleted mitochondrial GenomicAlignments}
\usage{
getMT(bam, chrM = "chrM", mtGenome = "hg19", plotMAPQ = FALSE)
}
\arguments{
\item{bam}{a BAM filename, which should have an index (else index it)}

\item{chrM}{what the mitochondrial contig is called. Default is "chrM"}

\item{mtGenome}{what mitochondrial assembly was used (default is hg19)}

\item{plotMAPQ}{plot distribution of mitochondrial mapping quality? (FALSE)}
}
\description{
grab the mitochondrial reads from a BAM and estimate their fraction
nb. this could probably be done faster for a list of BAMs but it's not
nb. nb. returns NuMt-depleted mitochondrial GenomicAlignments
}
