% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAlignments.R
\name{MAlignments}
\alias{MAlignments}
\title{wrap a GAlignments for easier stats}
\usage{
MAlignments(gal, readLength = 75, genomeSize = 16571)
}
\arguments{
\item{gal}{a GAlignments}

\item{readLength}{a number (read length)}

\item{genomeSize}{a number (number of bases in the host genome)}
}
\value{
an MAlignments
}
\description{
wrap a GAlignments for easier stats
}
