% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVRangesList.R
\docType{methods}
\name{MVRangesList-methods}
\alias{MVRangesList-methods}
\alias{coverage,MVRangesList-method}
\alias{counts<-,MVRangesList,RangedSummarizedExperiment-method}
\alias{counts,MVRangesList-method}
\alias{encoding,MVRangesList-method}
\alias{show,MVRangesList-method}
\alias{granges,MVRangesList-method}
\alias{summarizeVariants,MVRangesList,missing,missing-method}
\title{MVRangesList methods (centralized).}
\usage{
\S4method{coverage}{MVRangesList}(x)

\S4method{counts}{MVRangesList,RangedSummarizedExperiment}(object) <- value

\S4method{counts}{MVRangesList}(object)

\S4method{encoding}{MVRangesList}(x)

\S4method{show}{MVRangesList}(object)

\S4method{granges}{MVRangesList}(x, filterLowQual = TRUE)

\S4method{summarizeVariants}{MVRangesList,missing,missing}(query, subject, mode,
  ...)
}
\arguments{
\item{x}{an MVRangesList (for some methods)}

\item{object}{an MVRangesList (for other methods)}

\item{value}{a RangedSummarizedExperiment with matching colnames}

\item{filterLowQual}{optional argument to `granges` and `summarizeVariants`}

\item{annotations}{a RangedSummarizedExperiment with motif count matches}
}
\description{
`counts`               returns fragment counts, if any
`counts<-`             adds or updates fragment counts
`coverage`             returns estimated coverage for each element
`encoding`             returns mutations in coding regions for each element
`granges`              returns mildly annotated aggregates of variant sites
`summarizeVariants`    attempts mass functional annotation of variant sites
}
