% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVRangesList.R
\docType{methods}
\name{MVRangesList-methods}
\alias{MVRangesList-methods}
\alias{coverage,MVRangesList-method}
\alias{counts<-,MVRangesList,RangedSummarizedExperiment-method}
\alias{counts,MVRangesList-method}
\alias{encoding,MVRangesList-method}
\alias{show,MVRangesList-method}
\title{MVRangesList methods (centralized).}
\usage{
\S4method{coverage}{MVRangesList}(x)

\S4method{counts}{MVRangesList,RangedSummarizedExperiment}(object) <- value

\S4method{counts}{MVRangesList}(object)

\S4method{encoding}{MVRangesList}(x)

\S4method{show}{MVRangesList}(object)
}
\arguments{
\item{x}{an MVRangesList (for some methods)}

\item{object}{an MVRangesList (for other methods)}

\item{value}{a RangedSummarizedExperiment with matching colnames}

\item{annotations}{a RangedSummarizedExperiment with motif matches}
}
\description{
`counts`     returns fragment counts, if any, as a SummarizedExperiment.
`counts<-`   adds or updates fragment counts from a SummarizedExperiment.
`coverage`   returns the estimated coverage for each element in the list.
`deviations` returns deviations computed and stored from computeDeviations.
`encoding`   returns a subset of mutations in coding regions of mtDNA genes.
}
