% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAlignments.R
\docType{methods}
\name{runLength,MAlignments-method}
\alias{runLength,MAlignments-method}
\title{median read length (cached when possible)}
\usage{
\S4method{runLength}{MAlignments}(x)
}
\arguments{
\item{x}{an MAlignments}
}
\value{
estimated coverage (numeric)
}
\description{
median read length (cached when possible)
}
